+++ src/egon/data/db.py
@@ -204,10 +204,13 @@ def select_geodataframe(sql, index_col=None, geom_col="geom", epsg=3035):
 
     gdf = gpd.read_postgis(
         sql, engine(), index_col=index_col, geom_col=geom_col
-    ).to_crs(epsg=epsg)
+    )
 
     if gdf.size == 0:
         print(f"WARNING: No data returned by statement: \n {sql}")
+
+    else:
+        gdf = gdf.to_crs(epsg=epsg)
 
     return gdf
 
@@ -226,9 +229,15 @@ def next_etrago_id(component):
         Next index value
 
     """
+
+    if component=='transformer':
+        id_column = 'trafo_id'
+    else:
+        id_column = f'{component}_id'
+
     max_id = select_dataframe(
         f"""
-        SELECT MAX({component}_id) FROM grid.egon_etrago_{component}
+        SELECT MAX({id_column}) FROM grid.egon_etrago_{component}
         """
     )["max"][0]